requirements.txt
pyqt5==5.15.10
pyqt5-sip==12.13.0
pyqtwebengine==5.15.6
sqlalchemy==2.0.25
bcrypt==4.1.2
pyjwt==2.8.0
pyqterminal==0.1.0
reportlab==4.0.9
jinja2==3.1.3
markdown==3.5.2
pydantic==2.5.3
python-dotenv==1.0.1
